001_TablePagination/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap4==2.0.1
django-filter==2.3.0
django-tables2==2.3.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
